requirements.txt
captum.egg==info
librosa==0.10.1
matplotlib==3.8.4
numpy==2.0.1
pandas==2.0.3
Pillow==10.4.0
ptflops==0.7.2.2
scikit_learn==1.4.2
scipy==1.14.0
sympy==1.12
thop==0.1.1.post2209072238
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchinfo==1.8.0
torchvision==0.14.1+cu117
tqdm==4.66.2